backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31

django-cors-headers>=4.4
